Load the sms and email templates at startup to memory and reuse
